Add: Authentication token and driver for Login and Registration implemented

+++ EcoScrapbookingAPI.Presentation/Controllers/UserController.cs
@@ -1,10 +1,13 @@
+using System.Security.Claims;
 using EcoScrapbookingAPI.Business.DTOs.UserDTOs;
 using EcoScrapbookingAPI.Business.Interfaces;
 using EcoScrapbookingAPI.Domain.Models;
+using EcoScrapbookingAPI.Presentation.Utils;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace EcoScrapbookingAPI.Presentation.Controllers;
-
+[Authorize]
 [ApiController]
 [Route("[controller]")]
 public class UserController : ControllerBase
@@ -16,6 +19,7 @@ public class UserController : ControllerBase
     _userService = userService;
   }
 
+  [Authorize (Roles = "Admin")]
   [HttpGet]
   public ActionResult<List<UserGetDTO>> GetAll()
   {
@@ -34,11 +38,16 @@ public class UserController : ControllerBase
     }
   }
 
+  [Authorize]
   [HttpGet("{userId}")]
   public ActionResult<UserGetDTO> Get(int userId)
   {
     try
     {
+      if(!ControlUserAccess.UserHasAccess(User.FindFirst(ClaimTypes.Role).Value, User.FindFirst(ClaimTypes.NameIdentifier).Value, userId))
+      {
+        return Unauthorized("You do not have access to this user.");
+      }
       return Ok(_userService.GetUser(userId));
     }
     catch (ArgumentNullException anEx)
@@ -51,29 +60,16 @@ public class UserController : ControllerBase
     }
   }
 
-  [HttpPost]
-  public ActionResult<User> Create([FromBody] UserCreateDTO userCreateDTO)
-  {
-    try
-    {
-      var user = _userService.CreateUser(userCreateDTO);
-      return CreatedAtAction(nameof(Get), new { userId = user.UserId }, user);
-    }
-    catch (ArgumentNullException anEx)
-    {
-      return NotFound(anEx.Message);
-    }
-    catch (Exception ex)
-    {
-      return BadRequest(ex.Message);
-    }
-  }
-
+  [Authorize]
   [HttpPut("{userId}")]
   public ActionResult Update(int userId, [FromBody] UserUpdateDTO userUpdateDTO)
   {
     try
     {
+      if(!ControlUserAccess.UserHasAccess(User.FindFirst(ClaimTypes.Role).Value, User.FindFirst(ClaimTypes.NameIdentifier).Value, userId))
+      {
+        return Unauthorized("You do not have access to update this user.");
+      }
       _userService.UpdateUser(userId, userUpdateDTO);
       return Ok(userUpdateDTO);
     }
@@ -92,6 +88,10 @@ public class UserController : ControllerBase
   {
     try
     {
+      if(!ControlUserAccess.UserHasAccess(User.FindFirst(ClaimTypes.Role).Value, User.FindFirst(ClaimTypes.NameIdentifier).Value, userId))
+      {
+        return Unauthorized("You do not have access to delete this user.");
+      }
       _userService.DeleteUser(userId);
       return NoContent();
     }
@@ -123,11 +123,13 @@ public class UserController : ControllerBase
     }
   }
 
+  
   [HttpDelete("{userId}/activities/{activityId}")]
   public ActionResult RemoveUserFromActivity(int userId, int activityId)
   {
     try
     {
+      
       _userService.RemoveUserFromActivity(userId, activityId);
       return Ok();
     }
